requirements.txt
numpy==1.25.2
pandas==2.0.3
protobuf==4.23.4
python-dotenv==1.0.0
requests==2.31.0
google-cloud-bigquery==3.11.4
pyarrow==12.0.1
functions-framework==3.4.0
